Debug solver to working state
Solver is now working thanks to debugging with the tests. I missed a
line on the string input of the sudoku initialisation, so it was only
repeatedly setting the first square of the sudoku, now fixed. Also with
adding the values to the solution there was an off by one error, now
also fixed, and the infinite loop was caused by looping through the rows
of the node rather than the columns, that is now also fixed. It works!
Time to implement the GUI integration where a sudoku can be entered and
a solution can be found.