Add buttons to solver
Add solve and reset buttons to the gui solver. The solver now can take
input from the current state of the sudoku, solve it, and display that.
It can also reset it back to empty.

There is a bug with the way it works in that the table won't be updated
with the solution until the rows are clicked through, this is likely a
way that the JTable works. In essence it functions but the experience is
not super nice.